+++ examples/nextjs-comments-private/liveblocks.config.ts
@@ -1,0 +1,16 @@
+declare global {
+  interface Liveblocks {
+    // Custom user info set when authenticating with a secret key
+    UserMeta: {
+      id: string;
+      info: {
+        // Example properties, for useSelf, useUser, useOthers, etc.
+        name: string;
+        avatar: string;
+        color: string;
+      };
+    };
+  }
+}
+
+export {};